# internal/utils/error_response.go
package utils

import (
	"encoding/json"
	"log"
	"net/http"
)

type ErrorResponse struct {
	Error   string            `json:"error"`
	Details map[string]string `json:"details,omitempty"`
}

func JSONError(w http.ResponseWriter, massage string, statusCode int, details map[string]string) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(statusCode)
	response := ErrorResponse{
		Error:   massage,
		Details: details,
	}
	if err := json.NewEncoder(w).Encode(response); err != nil {
		log.Printf("Failed to send JSON error response: %v", err)
	}
}

func JSONSuccess(w http.ResponseWriter, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	if err := json.NewEncoder(w).Encode(data); err != nil {
		log.Printf("Failed to send JSON success response: %v", err)
	}
}